feat(api/flask/security): add basic authorization
User resources are protected behind a few authorization rules.
FeatureRequests are still open to any authenticated user.

closes #19

+++ featuring-flask-api/source/featuring/user/resources.py
@@ -2,9 +2,11 @@ import featuring
 from peewee import IntegrityError
 from flask_restful import Resource, abort, marshal_with
 from .models import User
+from featuring.security import jwt_protected, authorize
+from featuring.security.rules import IsAdmin
 from .serializers import UserSerializer
 from .parsers import UserCreateParser, UserPatchParser
-from featuring.security import jwt_protected
+from .rules import IsAdminOrSelf
 
 
 def find_or_abort(username):
@@ -20,9 +22,10 @@ class UserResource(Resource):
     @marshal_with(UserSerializer)
     def get(self, username):
         obj = find_or_abort(username)
-        return obj, 200
+        return authorize(IsAdminOrSelf, obj), 200
 
     def delete(self, username):
+        authorize(IsAdmin)
         obj = find_or_abort(username)
         obj.delete_instance()
         return '', 204
@@ -31,6 +34,7 @@ class UserResource(Resource):
     def patch(self, username):
         args = UserPatchParser.parse_args()
         obj = find_or_abort(username)
+        authorize(IsAdminOrSelf, obj)
         User.update(**args).where(User.username == username).execute()
         obj = find_or_abort(username)
         return obj, 201
@@ -41,11 +45,12 @@ class UserListResource(Resource):
 
     @marshal_with(UserSerializer)
     def get(self):
-        return list(User.select()
-            .order_by(User.username.asc()))
+        data = list(User.select().order_by(User.username.asc()))
+        return authorize(IsAdminOrSelf, data), 200
 
     @marshal_with(UserSerializer)
     def post(self):
+        authorize(IsAdmin)
         args = UserCreateParser.parse_args()
         created = User.create(**args)
         return created, 201
